+++ force_torque_sensor_broadcaster/test/test_load_force_torque_sensor_broadcaster.cpp
@@ -32,13 +32,17 @@ TEST(TestLoadForceTorqueSensorBroadcaster, load_controller)
     std::make_shared<rclcpp::executors::SingleThreadedExecutor>();
 
   controller_manager::ControllerManager cm(
-    executor, ros2_control_test_assets::minimal_robot_urdf, "test_controller_manager");
+    executor, ros2_control_test_assets::minimal_robot_urdf, true, "test_controller_manager");
 
-  ASSERT_NE(
-    cm.load_controller(
-      "test_force_torque_sensor_broadcaster",
-      "force_torque_sensor_broadcaster/ForceTorqueSensorBroadcaster"),
-    nullptr);
+  const std::string test_file_path =
+    std::string(TEST_FILES_DIRECTORY) + "/force_torque_sensor_broadcaster_params.yaml";
+
+  cm.set_parameter({"test_force_torque_sensor_broadcaster.params_file", test_file_path});
+  cm.set_parameter(
+    {"test_force_torque_sensor_broadcaster.type",
+     "force_torque_sensor_broadcaster/ForceTorqueSensorBroadcaster"});
+
+  ASSERT_NE(cm.load_controller("test_force_torque_sensor_broadcaster"), nullptr);
 }
 
 int main(int argc, char ** argv)