+++ tricycle_steering_controller/test/test_tricycle_steering_controller.cpp
@@ -14,10 +14,8 @@
 
 #include "test_tricycle_steering_controller.hpp"
 
-#include <limits>
 #include <memory>
 #include <string>
-#include <utility>
 #include <vector>
 
 class TricycleSteeringControllerTest
@@ -83,9 +81,9 @@ TEST_F(TricycleSteeringControllerTest, check_exported_interfaces)
   {
     const std::string ref_itf_name =
       std::string(controller_->get_node()->get_name()) + "/" + joint_reference_interfaces_[i];
-    EXPECT_EQ(ref_if_conf[i].get_name(), ref_itf_name);
-    EXPECT_EQ(ref_if_conf[i].get_prefix_name(), controller_->get_node()->get_name());
-    EXPECT_EQ(ref_if_conf[i].get_interface_name(), joint_reference_interfaces_[i]);
+    EXPECT_EQ(ref_if_conf[i]->get_name(), ref_itf_name);
+    EXPECT_EQ(ref_if_conf[i]->get_prefix_name(), controller_->get_node()->get_name());
+    EXPECT_EQ(ref_if_conf[i]->get_interface_name(), joint_reference_interfaces_[i]);
   }
 }
 
@@ -241,7 +239,7 @@ TEST_F(TricycleSteeringControllerTest, receive_message_and_publish_updated_statu
   EXPECT_EQ(msg.steering_angle_command[0], 2.2);
 
   publish_commands();
-  ASSERT_TRUE(controller_->wait_for_commands(executor));
+  controller_->wait_for_commands(executor);
 
   ASSERT_EQ(
     controller_->update(rclcpp::Time(0, 0, RCL_ROS_TIME), rclcpp::Duration::from_seconds(0.01)),